+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 0.13.5 (unreleased)
 
 - Minor refactorings: less private variables which are actually public (through properties) anyway
+- Store XML-node at UpnpDevice/UpnpService/UpnpAction/UpnpAction.Argument/UpnpStateVariable
 
 
 0.13.4 (2018-12-07)